refactor: Remover analytics e combinar personalização com configurações
- Remover página de analytics completamente
- Combinar funcionalidades de personalização na página de configurações
- Atualizar rotas removendo analytics e personalização
- Atualizar navegação do dashboard
- Simplificar estrutura do sistema mantendo todas as funcionalidades
